add two new recipes and minor changes
Two new recipes where added ("Pesto Pasta" and "Tomato soup").

The main page was modified in order to display the three recipes as an
unordered list.

The first recipe (Healthy Coconut Oatmeal) was modified in order to
indicate the web page from which it was obtained and to put a link
to the main page at the bottom.